Split Unpacker to own file
This patch removes Unpacker class from
bancks2resources.py to your own file
for a clean code pratice and S.O.L.I.D
principles.
